comparing oracle kpis with same number of floats

+++ oracle4grid/core/graph/indicators.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import numpy as np
 
 
 SHORTEST = "shortest"
@@ -55,28 +56,28 @@ def check_indicators_order(donothing, best, best_no_ol, best_without_transitions
 
     if best_path_type == SHORTEST:
         # Check if best path better than donothing
-        if donothing < best:
+        if np.float32(donothing) < np.float32(best):
             check = False
             message = "Indicator '" + MSG_DONOTHING + "' is better than indicator '" + MSG_GAMERULES
         # Check if best without game rule better than best with game rules
-        if best < best_without_transitions:
+        if np.float32(best) < np.float32(best_without_transitions):
             check = False
             message = "Indicator '" + MSG_GAMERULES + "' is better than indicator '" + MSG_NOGAMERULE
         # Check if best path better than best without overload
-        if pd.notnull(best_no_ol) and best_no_ol < best:
+        if pd.notnull(best_no_ol) and np.float32(best_no_ol) < np.float32(best):
             check = False
             message = "Indicator '" + MSG_GAMERULES_NO_OL + "' is better than indicator '" + MSG_GAMERULES
     elif best_path_type == LONGEST:
         # Check if best path better than donothing
-        if donothing > best:
+        if np.float32(donothing) > np.float32(best):
             check = False
             message = "Indicator '" + MSG_DONOTHING + "' is better than indicator '" + MSG_GAMERULES + " which should not be possible"
         # Check if best without game rule better than best with game rules
-        if best > best_without_transitions:
+        if np.float32(best) > np.float32(best_without_transitions):
             check = False
             message = "Indicator '" + MSG_GAMERULES + "' is better than indicator '" + MSG_NOGAMERULE + " which should not be possible"
         # Check if best path better than best without overload
-        if pd.notnull(best_no_ol) and best_no_ol > best:
+        if pd.notnull(best_no_ol) and np.float32(best_no_ol) > np.float32(best):
             check = False
             message = "Indicator '" + MSG_GAMERULES_NO_OL + "' is better than indicator '" + MSG_GAMERULES
     return check, message
